refactor(min): moved out of extended directory
BREAKING CHANGE: `min` has been moved out of `extended` folder and under `operator` folder directly
