--- src/add/operator/extended/min.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {min} from '../../../operator/extended/min';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.min = min;
-
-export var _void: void;
